notif/notif/requirement.txt
adal==1.2.2
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
argon2-cffi==19.1.0
attrs==19.3.0
azure-common==1.1.24
azure-servicebus==0.50.2
Babel==2.8.0
billiard==3.3.0.23
black==19.10b0
celery==3.1.26.post2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.4
django-allauth==0.39.1
django-appconf==1.0.3
django-autocomplete-light==3.4.1
django-celery==3.3.1
django-celery-beat==1.5.0
django-ckeditor==5.7.1
django-compressor==2.3
django-cors-headers==2.0.2
django-crispy-forms==1.7.2
django-environ==0.4.5
django-filter==2.2.0
django-import-export==1.2.0
django-js-asset==1.2.2
django-model-utils==3.2.0
django-notifications-hq==1.5.0
django-redis==4.10.0
django-timezone-field==4.0
django-webpack-loader==0.6.0
djangorestframework==3.10.2
drf-yasg==1.17.0
et-xmlfile==1.0.1
flower==0.9.3
idna==2.8
importlib-metadata==1.3.0
inflection==0.3.1
isodate==0.6.0
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.3
jsonfield==2.0.2
kombu==3.0.37
Markdown==3.1.1
MarkupPy==1.14
MarkupSafe==1.1.1
more-itertools==8.0.2
msrest==0.6.10
msrestazure==0.6.2
numpy==1.17.4
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==3.0.2
packaging==20.0
pathspec==0.7.0
pika==1.1.0
Pillow==6.1.0
pycparser==2.19
PyJWT==1.7.1
pyparsing==2.4.6
python-crontab==2.4.0
python-dateutil==2.8.1
python-slugify==3.0.3
python3-openid==3.1.0
pytz==2019.2
PyYAML==5.3
rcssmin==1.0.6
redis==3.3.6
regex==2020.1.8
requests==2.22.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
six==1.13.0
sqlparse==0.3.0
tablib==0.14.0
text-unidecode==1.2
toml==0.10.0
tornado==5.1.1
typed-ast==1.4.0
uamqp==1.2.5
uritemplate==3.0.1
urllib3==1.25.7
vine==1.3.0
whitenoise==4.1.3
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
django-json-field==0.5.7
